+++ app/pages/sections/additional-services/select/routes.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import { logger } from '../../../../logger';
 import config from '../../../../config';
-import { withCatch, extractAccessToken } from '../../../../helpers/routerHelper';
+import { withCatch, extractAccessToken } from '../../../../helpers/routes/routerHelper';
 import {
   findAdditionalServices,
   findAddedCatalogueSolutions,
