+++ app/pages/sections/supplier/routes.js
@@ -2,7 +2,7 @@ import express from 'express';
 import { ErrorContext } from 'buying-catalogue-library';
 import { logger } from '../../../logger';
 import config from '../../../config';
-import { withCatch, extractAccessToken } from '../../../helpers/routerHelper';
+import { withCatch, extractAccessToken } from '../../../helpers/routes/routerHelper';
 import {
   getSupplierSearchPageContext,
   validateSupplierSearchForm,
